Extract the event consuming part of UserClientRequestStrategy into UserClientEventConsumer
